More enemy spawn in rooms
