+++ core/src/com/dokkaebistudio/tacticaljourney/room/generation/floor2/Floor2RoomGenerator.java
@@ -3,21 +3,19 @@
  */
 package com.dokkaebistudio.tacticaljourney.room.generation.floor2;
 
-import java.util.ArrayList;
-import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
 
 import com.badlogic.ashley.core.Entity;
 import com.badlogic.gdx.math.RandomXS128;
 import com.badlogic.gdx.math.Vector2;
-import com.dokkaebistudio.tacticaljourney.Assets;
 import com.dokkaebistudio.tacticaljourney.ai.random.RandomSingleton;
-import com.dokkaebistudio.tacticaljourney.constants.ZIndexConstants;
+import com.dokkaebistudio.tacticaljourney.components.loot.LootRewardComponent;
 import com.dokkaebistudio.tacticaljourney.factory.EntityFactory;
-import com.dokkaebistudio.tacticaljourney.factory.EntityFlagEnum;
 import com.dokkaebistudio.tacticaljourney.room.Room;
 import com.dokkaebistudio.tacticaljourney.room.generation.GeneratedRoom;
 import com.dokkaebistudio.tacticaljourney.room.generation.RoomGenerator;
+import com.dokkaebistudio.tacticaljourney.util.Mappers;
 import com.dokkaebistudio.tacticaljourney.util.PoolableVector2;
 
 /**
@@ -79,4 +77,45 @@ public class Floor2RoomGenerator extends RoomGenerator {
 		generatedRoom.releaseSpawns();
 	}
 
+	
+	/**
+	 * Generate a random number of enemies and place them in a room.
+	 * @param room the room
+	 * @param random the random
+	 * @param spawnPositions the possible spawn positions
+	 * @param canBeEmpty true if there can be no enemies
+	 */
+	protected void placeEnemies(Room room, RandomXS128 random, List<PoolableVector2> spawnPositions, boolean canBeEmpty) {
+		if (spawnPositions.size() == 0) return;
+
+		int enemyNb = random.nextInt(Math.min(spawnPositions.size(), 8));
+		if (enemyNb == 0 && !canBeEmpty) enemyNb = 1;
+		
+		Iterator<PoolableVector2> iterator = spawnPositions.iterator();
+		for (int i=0 ; i<enemyNb ; i++) {
+			if (!iterator.hasNext()) break;
+			
+			Entity enemy = null;
+			int enemyTypeRandom = random.nextInt(7);
+			if (enemyTypeRandom == 0) {
+				enemy = entityFactory.enemyFactory.createSpider(room, new Vector2(iterator.next()), 3);
+				iterator.remove();
+			} else if (enemyTypeRandom == 1 || enemyTypeRandom == 2) {
+				enemy = entityFactory.enemyFactory.createSpiderWeb(room, new Vector2(iterator.next()), 4);
+				iterator.remove();
+				if (iterator.hasNext()) {
+					enemy = entityFactory.enemyFactory.createSpider(room, new Vector2(iterator.next()), 3);
+				}
+			} else if (enemyTypeRandom == 3 || enemyTypeRandom == 4) {
+				enemy = entityFactory.enemyFactory.createStinger(room, new Vector2(iterator.next()), 3);
+				iterator.remove();
+			} else {
+				enemy = entityFactory.enemyFactory.createScorpion(room, new Vector2(iterator.next()), 4);
+				iterator.remove();
+			}
+			
+			LootRewardComponent lootRewardComponent = Mappers.lootRewardComponent.get(enemy);
+			lootRewardComponent.setDrop( generateEnemyLoot(lootRewardComponent.getItemPool(), lootRewardComponent.getDropRate()));
+		}
+	}
 }
